Move test helper classes to separate file.
